Create header fro TFInfo
